@@ -11,20 +11,21 @@
     }
   ],
   "require": {
-    "php": ">=7.0.0",
+    "php": ">=7.4",
     "psr/cache": "^1.0",
     "psr/log": "^1.0",
     "psr/http-message": "^1.0",
+    "ramsey/uuid": "^3.8",
     "ext-json": "*",
     "ext-filter": "*",
-    "ramsey/uuid": "^3.8"
+    "ext-mbstring": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "~6.0",
-    "squizlabs/php_codesniffer": "~3.0",
-    "php-coveralls/php-coveralls": "~1.0",
-    "phpstan/phpstan": "~0.10.2",
-    "maglnet/composer-require-checker": "~1.0"
+    "phpunit/phpunit": "^9.0",
+    "squizlabs/php_codesniffer": "^3.0",
+    "php-coveralls/php-coveralls": "^2.0",
+    "phpstan/phpstan": "^0.12.19",
+    "maglnet/composer-require-checker": "^2.1"
   },
   "autoload": {
     "psr-4": {
