@@ -1,34 +1,24 @@
 <?php
+
+declare(strict_types=1);
+
 namespace RemotelyLiving\Doorkeeper\Features;
 
-class SetFactory
+final class SetFactory
 {
-    /**
-     * @var \RemotelyLiving\Doorkeeper\Features\Factory
-     */
-    private $feature_factory;
+    private Factory $featureFactory;
 
-    /**
-     * SetFactory constructor.
-     *
-     * @param \RemotelyLiving\Doorkeeper\Features\Factory|null $factory
-     */
     public function __construct(Factory $factory = null)
     {
-        $this->feature_factory = $factory ?? new Factory();
+        $this->featureFactory = $factory ?? new Factory();
     }
 
-    /**
-     * @param array $feature_set
-     *
-     * @return \RemotelyLiving\Doorkeeper\Features\Set
-     */
-    public function createFromArray(array $feature_set): Set
+    public function createFromArray(array $featureSet): Set
     {
         $features = [];
 
-        foreach ($feature_set as $feature) {
-            $features[] = $this->feature_factory->createFromArray($feature);
+        foreach ($featureSet as $feature) {
+            $features[] = $this->featureFactory->createFromArray($feature);
         }
 
         return new Set($features);
