@@ -1,63 +1,66 @@
 <?php
+
+declare(strict_types=1);
+
 namespace RemotelyLiving\Doorkeeper\Rules;
 
 /**
  * A type mapper if you choose to store integer types for persisted rules
  */
-class TypeMapper
+final class TypeMapper
 {
-    const RULE_TYPE_HEADER          = 1;
-    const RULE_TYPE_IP_ADDRESS      = 2;
-    const RULE_TYPE_PERCENTAGE      = 3;
-    const RULE_TYPE_RANDOM          = 4;
-    const RULE_TYPE_STRING_HASH     = 5;
-    const RULE_TYPE_USER_ID         = 6;
-    const RULE_TYPE_ENVIRONMENT     = 7;
-    const RULE_TYPE_BEFORE          = 8;
-    const RULE_TYPE_AFTER           = 9;
-    const RULE_TYPE_PIPED_COMPOSITE = 10;
+    public const RULE_TYPE_HEADER = 1;
+    public const RULE_TYPE_IP_ADDRESS = 2;
+    public const RULE_TYPE_PERCENTAGE = 3;
+    public const RULE_TYPE_RANDOM = 4;
+    public const RULE_TYPE_STRING_HASH = 5;
+    public const RULE_TYPE_USER_ID = 6;
+    public const RULE_TYPE_ENVIRONMENT = 7;
+    public const RULE_TYPE_BEFORE = 8;
+    public const RULE_TYPE_AFTER = 9;
+    public const RULE_TYPE_PIPED_COMPOSITE = 10;
 
-    private $type_map = [
-        self::RULE_TYPE_HEADER          => HttpHeader::class,
-        self::RULE_TYPE_IP_ADDRESS      => IpAddress::class,
-        self::RULE_TYPE_PERCENTAGE      => Percentage::class,
-        self::RULE_TYPE_RANDOM          => Random::class,
-        self::RULE_TYPE_STRING_HASH     => StringHash::class,
-        self::RULE_TYPE_USER_ID         => UserId::class,
-        self::RULE_TYPE_ENVIRONMENT     => Environment::class,
-        self::RULE_TYPE_BEFORE          => TimeBefore::class,
-        self::RULE_TYPE_AFTER           => TimeAfter::class,
+    private array $typeMap = [
+        self::RULE_TYPE_HEADER => HttpHeader::class,
+        self::RULE_TYPE_IP_ADDRESS => IpAddress::class,
+        self::RULE_TYPE_PERCENTAGE => Percentage::class,
+        self::RULE_TYPE_RANDOM => Random::class,
+        self::RULE_TYPE_STRING_HASH => StringHash::class,
+        self::RULE_TYPE_USER_ID => UserId::class,
+        self::RULE_TYPE_ENVIRONMENT => Environment::class,
+        self::RULE_TYPE_BEFORE => TimeBefore::class,
+        self::RULE_TYPE_AFTER => TimeAfter::class,
         self::RULE_TYPE_PIPED_COMPOSITE => PipedComposite::class,
     ];
 
-    public function __construct(array $extra_types = [])
+    public function __construct(array $extraTypes = [])
     {
-        foreach ($extra_types as $id => $name) {
+        foreach ($extraTypes as $id => $name) {
             $this->pushExtraType($id, $name);
         }
     }
 
-    public function getClassNameById(int $integer_id): string
+    public function getClassNameById(int $integerId): string
     {
-        return $this->type_map[$integer_id];
+        return $this->typeMap[$integerId];
     }
 
-    public function getIdForClassName(string $class_name): int
+    public function getIdForClassName(string $className): int
     {
-        return array_flip($this->type_map)[$class_name];
+        return array_flip($this->typeMap)[$className];
     }
 
-    public function pushExtraType(int $id, string $class_name)
+    public function pushExtraType(int $id, string $className): void
     {
 
-        if (isset($this->type_map[$id])) {
+        if (isset($this->typeMap[$id])) {
             throw new \DomainException("Type {$id} already set by parent");
         }
 
-        if (!class_exists($class_name)) {
-            throw new \InvalidArgumentException("{$class_name} is not a loaded class");
+        if (!class_exists($className)) {
+            throw new \InvalidArgumentException("{$className} is not a loaded class");
         }
 
-        $this->type_map[$id] = $class_name;
+        $this->typeMap[$id] = $className;
     }
 }
