@@ -1,9 +1,12 @@
 <?php
+
+declare(strict_types=1);
+
 namespace RemotelyLiving\Doorkeeper\Identification;
 
-class IntegerId extends IdentificationAbstract
+final class IntegerId extends AbstractIdentification
 {
-    public function validate($id)
+    protected function validate($id): void
     {
         if (!is_int($id) || $id < 0) {
             throw new \InvalidArgumentException("{$id} is not a positive integer id");
