@@ -1,13 +1,18 @@
 <?php
+
+declare(strict_types=1);
+
 namespace RemotelyLiving\Doorkeeper\Identification;
 
-class IpAddress extends IdentificationAbstract
+final class IpAddress extends AbstractIdentification
 {
-    public function validate($ip_address)
+    protected function validate($ipAddress): void
     {
-        if (!filter_var($ip_address, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4)
-          && !filter_var($ip_address, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
-            throw new \InvalidArgumentException("{$ip_address} is not a well formed ip address");
+        if (
+            !filter_var($ipAddress, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4)
+            && !filter_var($ipAddress, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)
+        ) {
+            throw new \InvalidArgumentException("{$ipAddress} is not a well formed ip address");
         }
     }
 }
